Retry if the connection has been closed by the CouchDB server.
Mochiweb has a default idle timeout of 30s after which it closes the
connection. The first time we know this has happened is when we try to read the
response and httplib raises a BadStatusLine exception with an empty line.
Instead of failing, we now retry the request.

--HG--
branch : experimental/httplib