Update example python view code to avoid KeyErrors
If the reader doesn't happen to have the 'date' key in one of the documents the old code would cause couchdb to throw. If the reader happened to be using futon, it would produce a useless 'no response' error.

By using .get() you'll always at least get None and this code can be used against any database.