Make Client.update_doc documentation clearer
I was a little thrown off by the documentation here, as there didn't
seem to be a way to provide my own request body to the request when
calling an update handler; the documentation only indicates that query
string params can be given as extra arguments.

A little digging into the code shows that options are just passed to the
post/put methods intact, so params of those methods can indeed be
provided, including path (probably not recommended!), body, and headers.
I've updated the documentation accordingly.