Add microseconds support for DateTimeField (#293)
* Add microseconds support for DateTimeField

A check is introduced in the DateTimeField to understand if microseconds
are provided in the datetime string or not. The return value is then
handled in accordance to that check.

Tests related to the feature implemented are also introduced.

Doctest examples are updated too.

* Add microseconds support for DateTimeField

A check is introduced in the DateTimeField to understand if microseconds
are provided in the datetime string or not. The return value is then
handled in accordance to that check.

Tests related to the feature implemented are also introduced.

Doctest examples are updated too.